clean up and random implement
Cleaned up quite a few lines of where the game talked to the user making it seem more realistic as well as fixed minor typo's. i then implemented the random function so the game isn't so static there is now a random feature to the health's and attacks of both you and your enemy.